requirements.txt
django==2.2
psycopg2==2.8.3
djangorestframework==3.10.3
django-storages==1.7 
boto3==1.9
